add: add UserService

+++ src/Services/RedisService.ts
@@ -1,44 +1,40 @@
 import { ConfigurationService } from "./Configurations/ConfigurationService";
-import * as Redis from "redis"
+import {
+    RedisClientType,
+    createClient
+} from "redis"
 
 export class RedisService {
     private readonly _configurationService: ConfigurationService;
-    private client: Redis.RedisClientType;
+    private readonly client: RedisClientType;
+    private isConnected: boolean = false;
 
     constructor(
         configurationService: ConfigurationService,
     ) {
         this._configurationService = configurationService;
-        this.client = Redis.createClient({
+        this.client = createClient({
             socket: {
                 host: this._configurationService.REDIS_HOST,
                 port: this._configurationService.REDIS_PORT,
             },
-            password: this._configurationService.REDIS_PASSWORD,
         });
-    }
-    
-    public async isConnected(): Promise<boolean>{
-        const request = await this.client.ping();
-        return request === 'PONG' ? true : false;
+        this.client.on('ready', () => this.isConnected = true)
     }
 
-    public async setValue(key: string, value: string): Promise<string | null>{
-        return await this.client.set(key, value);
+
+
+    public async setValue(key: string, value: string): Promise<string | null> {
+        if (this.isConnected === true) return await this.client.set(key, value);
+        else return null;
     }
 
-    public async getValue(key: string): Promise<string | null>{
-        try{
-            return await this.client.get(key);
-        } catch{
-            return null;
-        }
+    public async getValue(key: string): Promise<string | null> {
+        if (this.isConnected === true) return await this.client.get(key);
+        else return null;
     }
-    public async delete(key: string): Promise<number | null>{
-        try{
-            return await this.client.del(key);
-        } catch{
-            return null;
-        }
+    public async remove(key: string): Promise<number | null> {
+        if (this.isConnected === true) return await this.client.del(key);
+        else return null;
     }
 }
